Moved .sln into source. Rename to EntryPoint.SubDomain. Rename to EntryPoint.MarketOperator.
