+++ source/Energinet.DataHub.PostOffice.EntryPoint.MarketOperator/Functions/Peek.cs
@@ -23,35 +23,36 @@ using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
 using Microsoft.Extensions.Logging;
 
-namespace Energinet.DataHub.PostOffice.Outbound.Functions
+namespace Energinet.DataHub.PostOffice.EntryPoint.MarketOperator.Functions
 {
-    public sealed class Dequeue
+    public sealed class Peek
     {
         private readonly IMediator _mediator;
 
-        public Dequeue(IMediator mediator)
+        public Peek(IMediator mediator)
         {
             _mediator = mediator;
         }
 
-        [Function("Dequeue")]
+        [Function("Peek")]
         public async Task<HttpResponseMessage> RunAsync(
-            [HttpTrigger(AuthorizationLevel.Function, "delete")]
+            [HttpTrigger(AuthorizationLevel.Function, "get")]
             HttpRequestData request,
             FunctionContext context)
         {
             if (request == null)
                 throw new ArgumentNullException(nameof(request));
 
-            var logger = context.GetLogger<Dequeue>();
-            var command = request.Url.ParseQuery<DequeueCommand>();
+            var logger = context.GetLogger<Peek>();
+            var command = request.Url.ParseQuery<PeekCommand>();
 
-            logger.LogInformation($"Processing Dequeue query: {command}.");
+            logger.LogInformation($"Processing Peek query: {command}.");
 
-            var response = await _mediator.Send(command).ConfigureAwait(false);
-            return response.IsDequeued
-                ? new HttpResponseMessage(HttpStatusCode.OK)
-                : new HttpResponseMessage(HttpStatusCode.NotFound);
+            var (hasContent, stream) = await _mediator.Send(command).ConfigureAwait(false);
+
+            return hasContent
+                ? new HttpResponseMessage(HttpStatusCode.OK) { Content = new StreamContent(stream) }
+                : new HttpResponseMessage(HttpStatusCode.NoContent);
         }
     }
 }
