+++ source/Energinet.DataHub.PostOffice.Infrastructure/Documents/BundleDocument.cs
@@ -12,14 +12,9 @@
 // // See the License for the specific language governing permissions and
 // // limitations under the License.
 
-using System;
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
-using System.Linq;
-using System.Text.Json.Serialization;
-using Energinet.DataHub.PostOffice.Domain.Model;
 
-namespace Energinet.DataHub.PostOffice.Infrastructure.Entities
+namespace Energinet.DataHub.PostOffice.Infrastructure.Documents
 {
     internal record BundleDocument
     {
