+++ source/Energinet.DataHub.PostOffice.sln
@@ -13,27 +13,27 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # Copyright 2020 Energinet DataHub A/S
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.PostOffice.Domain", "source\Energinet.DataHub.PostOffice.Domain\Energinet.DataHub.PostOffice.Domain.csproj", "{A8D6B689-482A-4964-9B3C-1BF9DA9496E3}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.PostOffice.Domain", "Energinet.DataHub.PostOffice.Domain\Energinet.DataHub.PostOffice.Domain.csproj", "{A8D6B689-482A-4964-9B3C-1BF9DA9496E3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.PostOffice.Application", "source\Energinet.DataHub.PostOffice.Application\Energinet.DataHub.PostOffice.Application.csproj", "{B3CC1EFC-DA6C-45F3-9AA6-D75FDBE1504B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.PostOffice.Application", "Energinet.DataHub.PostOffice.Application\Energinet.DataHub.PostOffice.Application.csproj", "{B3CC1EFC-DA6C-45F3-9AA6-D75FDBE1504B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.PostOffice.Infrastructure", "source\Energinet.DataHub.PostOffice.Infrastructure\Energinet.DataHub.PostOffice.Infrastructure.csproj", "{4EA05C50-C0AE-4ACB-90BA-A4433A9A1C61}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.PostOffice.Infrastructure", "Energinet.DataHub.PostOffice.Infrastructure\Energinet.DataHub.PostOffice.Infrastructure.csproj", "{4EA05C50-C0AE-4ACB-90BA-A4433A9A1C61}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.PostOffice.Tests", "source\Energinet.DataHub.PostOffice.Tests\Energinet.DataHub.PostOffice.Tests.csproj", "{68412EBF-701D-467D-8716-74269E171F41}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.PostOffice.Tests", "Energinet.DataHub.PostOffice.Tests\Energinet.DataHub.PostOffice.Tests.csproj", "{68412EBF-701D-467D-8716-74269E171F41}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Hosts", "Hosts", "{832447E5-7A46-4381-817C-5D73A20510C2}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{05DE980B-A2A7-4235-AE53-3167E94699DD}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.PostOffice.Outbound", "source\Energinet.DataHub.PostOffice.Outbound\Energinet.DataHub.PostOffice.Outbound.csproj", "{02B328BB-2665-4933-81D1-AC4DCDDF5745}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.PostOffice.EntryPoint.MarketOperator", "Energinet.DataHub.PostOffice.EntryPoint.MarketOperator\Energinet.DataHub.PostOffice.EntryPoint.MarketOperator.csproj", "{02B328BB-2665-4933-81D1-AC4DCDDF5745}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.PostOffice.Inbound", "source\Energinet.DataHub.PostOffice.Inbound\Energinet.DataHub.PostOffice.Inbound.csproj", "{0F71324E-779C-4899-B995-B1227F8942A6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.PostOffice.EntryPoint.SubDomain", "Energinet.DataHub.PostOffice.EntryPoint.SubDomain\Energinet.DataHub.PostOffice.EntryPoint.SubDomain.csproj", "{0F71324E-779C-4899-B995-B1227F8942A6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.PostOffice.ServiceContracts", "source\Energinet.DataHub.PostOffice.ServiceContracts\Energinet.DataHub.PostOffice.ServiceContracts.csproj", "{BB4BCD78-1058-4D5A-90D4-82EF4BBFDCAC}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.PostOffice.ServiceContracts", "Energinet.DataHub.PostOffice.ServiceContracts\Energinet.DataHub.PostOffice.ServiceContracts.csproj", "{BB4BCD78-1058-4D5A-90D4-82EF4BBFDCAC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.PostOffice.Common", "source\Energinet.DataHub.PostOffice.Common\Energinet.DataHub.PostOffice.Common.csproj", "{6265B395-DFFE-45BA-899E-97F473115E2B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.PostOffice.Common", "Energinet.DataHub.PostOffice.Common\Energinet.DataHub.PostOffice.Common.csproj", "{6265B395-DFFE-45BA-899E-97F473115E2B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.PostOffice.IntegrationTests", "source\Energinet.DataHub.PostOffice.IntegrationTests\Energinet.DataHub.PostOffice.IntegrationTests.csproj", "{4EE75C4A-6EA6-4438-875D-040B8296A3D3}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.PostOffice.IntegrationTests", "Energinet.DataHub.PostOffice.IntegrationTests\Energinet.DataHub.PostOffice.IntegrationTests.csproj", "{4EE75C4A-6EA6-4438-875D-040B8296A3D3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
